Rename `tutorial-panel` folder to `tutorial`
Since it will also house UI components for the "Welcome" and "You're Done!" panes
