+++ app/src/ui/tutorial/tutorial-panel.tsx
